+++ bioconda_utils/update.py
@@ -49,37 +49,31 @@ import logging
 import os
 import pickle
 import random
-import re
 
 from collections import defaultdict, Counter
 from concurrent.futures import ProcessPoolExecutor
-from copy import copy
+
 from hashlib import sha256
 from urllib.parse import urlparse
-from typing import Any, Dict, List, Mapping, Optional, Sequence, Pattern
+from typing import Any, Dict, List, Mapping, Optional, Sequence, Set, Tuple, Type
 
 import aiohttp
 import aiofiles
 import aioftp
 import backoff
+import yaml
 
 import conda_build.variants
 import conda_build.config
 from conda.exports import MatchSpec, VersionOrder
 import conda.exceptions
 
-try:
-    from ruamel.yaml import YAML
-    from ruamel.yaml.constructor import DuplicateKeyError
-except ModuleNotFoundError:
-    from ruamel_yaml import YAML
-    from ruamel_yaml.constructor import DuplicateKeyError
-
 from pkg_resources import parse_version
 
 from . import utils
 from .utils import ensure_list, tqdm
 from .githubhandler import AiohttpGitHubHandler, GitHubHandler
+from .recipe import Recipe, RecipeError
 
 # pkg_resources.parse_version returns a Version or LegacyVersion object
 # as defined in packaging.version. Since it's bundling it's own copy of
@@ -87,13 +81,6 @@ from .githubhandler import AiohttpGitHubHandler, GitHubHandler
 # So we cheat by having it create a LegacyVersion delibarately.
 LegacyVersion = parse_version("").__class__  # pylint: disable=invalid-name
 
-yaml = YAML(typ="rt")  # pylint: disable=invalid-name
-
-# Hack: mirror stringify from conda-build in removing implicit
-#       resolution of numbers
-for digit in '0123456789':
-    if digit in yaml.resolver.versioned_resolver:
-        del yaml.resolver.versioned_resolver[digit]
 
 logger = logging.getLogger(__name__)  # pylint: disable=invalid-name
 
@@ -102,430 +89,9 @@ class EndProcessing(BaseException):
     """Raised by `Filter` to tell `Scanner` to stop processing"""
 
 
-class RecipeError(Exception):
-    """Raised to indicate processing of Recipe failed"""
-    __slots__ = ['recipe', 'args']
-    template = "broken: %s"
-    level = logging.INFO
-
-    def __init__(self, recipe: "Recipe", *args) -> None:
-        super().__init__(recipe, *args)
-        self.recipe = recipe
-        self.args = args
-
-    def log(self, uselogger=logger, level=None):
-        """Print message using provided logging func"""
-        if not level:
-            level = self.level
-        uselogger.log(level, "Recipe %s " + self.template, self.recipe, *self.args)
-
-    @property
-    def name(self):
-        """Class name of RecipeError"""
-        return self.__class__.__name__
-
-
-class Recipe():
-    """Represents a recipe (meta.yaml) in editable form
-
-    Using conda-build to render recipe is slow and a one-way
-    process. We need to be able to load **and** save recipes, which is
-    handled by the representation in this class.
-
-    Recipes undergo two manipulation rounds before parsed as YAML:
-     1. Selecting lines using ``# [expression]``
-     2. Rendering as Jinja2 template
-
-    (1) is currently unhandled, leading to recipes with repeated mapping keys
-    (commonly two ``url`` keys). Those recipes are ignored for the time being.
-
-    Arguments:
-      recipe_folder: base recipes folder
-      recipe_dir: path to specific recipe
-    """
-
-    class DuplicateKey(RecipeError):
-        """Raised for recipes with duplicate keys in the meta.yaml. YAML
-        does not allow those, but the PyYAML parser silently overwrites
-        previous keys.
-
-        For duplicate keys that are a result of `# [osx]` style line selectors,
-        `Recipe` attempts to resolve them as a list of dictionaries instead.
-        """
-        template = "has duplicate key"
-
-    class MissingKey(RecipeError):
-        """Raised if a recipe is missing package/version or package/name"""
-        template = "has missing key"
-
-    class EmptyRecipe(RecipeError):
-        """Raised if the recipe file is empty"""
-        template = "is empty"
-
-    class MissingBuild(RecipeError):
-        """Raised if the recipe is missing the build section"""
-        template = "is missing build section"
-
-    class HasSelector(RecipeError):
-        """Raised when recplacements fail due to `# [cond]` line selectors
-        FIXME: This should no longer be an error
-        """
-        template = "has selector in line %i (replace failed)"
-
-    #: Variables to pass to Jinja when rendering recipe
-    JINJA_VARS = {
-        "cran_mirror": "https://cloud.r-project.org",
-        "compiler": lambda x: f"compiler_{x}",
-    }
-
-    #: Name of key under ``extra`` containing config
-    EXTRA_CONFIG = "autobump"
-
-    def __init__(self, recipe_dir, recipe_folder):
-        if not recipe_dir.startswith(recipe_folder):
-            raise RuntimeError(f"'{recipe_dir}' not inside '{recipe_folder}'")
-
-        #: path to folder containing recipes
-        self.basedir = recipe_folder
-        #: relative path to recipe dir from folder containing recipes
-        self.reldir = recipe_dir[len(recipe_folder):].strip("/")
-
-        # Filled in by render()
-        #: Parsed recipe YAML
-        self.meta: Dict[str, Any] = {}
-
-        # These will be filled in by load_from_string()
-        #: Lines of the raw recipe file
-        self.meta_yaml: List[str] = []
-        # Filled in by update filter
-        self.version_data = None
-        #: Original recipe before modifications (updated by load_from_string)
-        self.orig: Recipe = copy(self)
-        #: Whether the recipe was loaded from a branch (update in progress)
-        self.on_branch: bool = False
-
-    @property
-    def path(self):
-        """Full path to `meta.yaml``"""
-        return os.path.join(self.basedir, self.reldir, "meta.yaml")
-
-    @property
-    def config(self):
-        """Per-recipe configuration parameters
-
-        These are the values set in ``extra:`` under the key
-        defined as `Recipe.EXTRA_CONFIG` (default is ``watch``).
-        """
-        return self.meta.get("extra", {}).get(self.EXTRA_CONFIG, {})
-
-    def __str__(self) -> str:
-        return self.reldir
-
-    def load_from_string(self, data) -> "Recipe":
-        """Load and `render` recipe contents from disk"""
-        self.meta_yaml = data.splitlines()
-        if not self.meta_yaml:
-            raise self.EmptyRecipe(self)
-        self.render()
-        return self
-
-    def set_original(self) -> None:
-        """Store the current state of the recipe as "original" version"""
-        self.orig = copy(self)
-
-    def dump(self):
-        """Dump recipe content"""
-        return "\n".join(self.meta_yaml) + "\n"
-
-    @staticmethod
-    def _rewrite_selector_block(text, block_top, block_left):
-        if not block_left:
-            return None  # never the whole yaml
-        lines = text.splitlines()
-        block_height = None
-        variants = defaultdict(list)
-
-        for block_height, line in enumerate(lines[block_top:]):
-            if line.strip() and not line.startswith(" " * block_left):
-                break
-            _, _, selector = line.partition("#")
-            if selector:
-                variants[selector.strip("[] ")].append(line)
-            else:
-                for variant in variants:
-                    variants[variant].append(line)
-        else:
-            # end of file, need to add one to block height
-            block_height+=1
-
-        if not block_height:  # empty lines?
-            return None
-        if not variants:
-            return None
-        if any(" " in v for v in variants):
-            # can't handle "[py2k or osx]" style things
-            return None
-
-        new_lines = []
-        for variant in variants.values():
-            first = True
-            for line in variant:
-                if first:
-                    new_lines.append("".join((" " * block_left, "- ", line)))
-                    first = False
-                else:
-                    new_lines.append("".join((" " * (block_left + 2), line)))
-
-        logger.debug("Replacing: lines %i - %i with %i lines:\n%s\n---\n%s",
-                     block_top, block_top+block_height, len(new_lines),
-                     "\n".join(lines[block_top:block_top+block_height]),
-                     "\n".join(new_lines))
-
-        lines[block_top:block_top+block_height] = new_lines
-        return "\n".join(lines)
-
-    def get_template(self):
-        """Create a Jinja2 template from the current raw recipe"""
-        # This function exists because the template cannot be pickled.
-        # Storing it means the recipe cannot be pickled, which in turn
-        # means we cannot pass it to ProcessExecutors.
-        return utils.jinja_silent_undef.from_string(
-            "\n".join(self.meta_yaml)
-        )
-
-    def get_simple_modules(self):
-        """Yield simple replacement values from template
-
-        E.g. those set with ``{% set version='1.2.3' %}```
-
-        Returns:
-          a dictionary of replacements (e.g. version: 1.2.3)
-        """
-        template = self.get_template()
-        return {
-            attr: getattr(template.module, attr)
-            for attr in dir(template.module)
-            if not attr.startswith("_")
-            and not hasattr(getattr(template.module, attr), '__call__')
-        }
-
-    def render(self) -> None:
-        """Convert recipe text into data structure
-
-        - create jinja template from recipe content
-        - render template
-        - parse yaml
-        - normalize
-        """
-        yaml_text = self.get_template().render(self.JINJA_VARS)
-        try:
-            self.meta = yaml.load(yaml_text)
-        except DuplicateKeyError as err:
-            logger.debug("fixing duplicate key at %i:%i",
-                         err.context_mark.line, err.context_mark.column)
-            # We may have encountered a recipe with linux/osx variants using line selectors
-            yaml_text = self._rewrite_selector_block(yaml_text, err.context_mark.line,
-                                                     err.context_mark.column)
-            if yaml_text:
-                try:
-                    self.meta = yaml.load(yaml_text)
-                except DuplicateKeyError as errx:
-                    raise self.DuplicateKey(self)
-            else:
-                raise self.DuplicateKey(self)
-
-        if ("package" not in self.meta
-            or "version" not in self.meta["package"]
-            or "name" not in self.meta["package"]):
-            raise self.MissingKey(self)
-
-        # make recipe-maintainers a list if it is a string
-        maintainers = self.meta.mlget(["extra", "recipe-maintainers"])
-        if maintainers and isinstance(maintainers, str):
-            self.meta["extra"]["recipe-maintainers"] = [maintainers]
-
-    @property
-    def name(self) -> str:
-        """The name of the toplevel package built by this recipe"""
-        return self.meta["package"]["name"]
-
-    @property
-    def version(self) -> str:
-        """The version of the package build by this recipe"""
-        return str(self.meta["package"]["version"])
-
-    @property
-    def package_names(self) -> List[str]:
-        """List of the packages built by this recipe (including outputs)"""
-        packages = [self.name]
-        if "outputs" in self.meta:
-            packages.extend(output['name'] for output in self.meta['outputs'])
-        return packages
-
-    def replace(self, before: str, after: str,
-                within: Sequence[str] = ("package", "source"),
-                with_fuzz=False) -> int:
-        """Runs string replace on parts of recipe text.
-
-        - Lines considered are those containing Jinja set statements
-        (``{% set var="val" %}``) and those defining the top level
-        Mapping entries given by **within** (default:``package`` and
-        ``source``).
-
-        - Cowardly refuses to modify lines with ``# [expression]``
-        selectors.
-        """
-        logger.debug("Trying to replace %s with %s", before, after)
-
-        # get lines starting with "{%"
-        lines = set()
-        for lineno, line in enumerate(self.meta_yaml):
-            if line.strip().startswith("{%"):
-                lines.add(lineno)
-
-        # get lines covered by keys listed in ``within``
-        start: Optional[int] = None
-        for key in self.meta.keys():
-            lineno = self.meta.lc.key(key)[0]
-            if key in within:
-                if start is None:
-                    start = lineno
-            else:
-                if start is not None:
-                    lines.update(range(start, lineno))
-                    start = None
-        if start is not None:
-            lines.update(range(start, len(self.meta_yaml)))
-
-        if isinstance(before, Pattern):
-            re_before = before
-            re_select = re.compile(before.pattern + r".*#.*\[")
-        else:
-            before_pattern = re.escape(before)
-            if with_fuzz:
-                before_pattern = re.sub(r"(-|\\\.|_)", "[-_.]", before_pattern)
-            re_before = re.compile(before_pattern)
-            re_select = re.compile(before_pattern + r".*#.*\[")
-
-        # replace within those lines, erroring on "# [asd]" selectors
-        replacements = 0
-        for lineno in sorted(lines):
-            line = self.meta_yaml[lineno]
-            if not re_before.search(line):
-                continue
-            if re_select.search(line):
-                raise self.HasSelector(self, lineno)
-            new = re_before.sub(after, line)
-            logger.debug("%i - %s", lineno, self.meta_yaml[lineno])
-            logger.debug("%i + %s", lineno, new)
-            self.meta_yaml[lineno] = new
-            replacements += 1
-        return replacements
-
-    def reset_buildnumber(self):
-        """Resets the build number
-
-        If the build number is missing, it is added after build.
-        """
-        try:
-            lineno: int = self.meta["build"].lc.key("number")[0]
-        except (KeyError, AttributeError):  # no build number?
-            if "build" in self.meta and self.meta["build"] is not None:
-                build = self.meta["build"]
-                first_in_build = next(iter(build))
-                lineno, colno = build.lc.key(first_in_build)
-                self.meta_yaml.insert(lineno, " "*colno + "number: 0")
-            else:
-                raise self.MissingBuild(self)
-
-        line = self.meta_yaml[lineno]
-        line = re.sub("number: [0-9]+", "number: 0", line)
-        self.meta_yaml[lineno] = line
-
-    def get_raw_range(self, path):
-        """Locate the position of a node in the YAML within the raw text
-
-
-        See also `get_raw()` if you want to get the content of the unparsed
-        meta.yaml at a specific key.
-
-        Args:
-          path: The "path" to the node. Use numbers for lists ('source/1/url')
-
-        Returns:
-          a tuple of first_row, first_column, last_row, last_column
-        """
-        nodes = [self.meta]
-        keys = []
-        # find items composing path
-        for key in path.split("/"):
-            last = nodes[-1]
-            try:
-                number = int(key)
-                if isinstance(last, list):
-                    nodes.append(last[number])
-                    keys.append(number)
-                elif isinstance(last, dict) and number == 0:
-                    pass
-                else:
-                    raise ValueError(f"Can't access {path} in meta")
-            except ValueError:
-                nodes.append(last[key])
-                keys.append(key)
-
-        nodes.pop()  # pop parsed value
-
-        # get the start row/col for the value
-        if isinstance(keys[-1], int):
-            start_row, start_col = nodes[-1].lc.key(keys[-1])
-        else:
-            start_row, start_col = nodes[-1].lc.value(keys[-1])
-
-        # getting the end is more complicated, we need to move
-        # up the tree to the next item in order until one is not the last
-        # item in its collection
-        while nodes:
-            node = nodes.pop()
-            key = keys.pop()
-            if isinstance(key, int):
-                if key + 1 < len(node):
-                    end_row, end_col = node.lc.key(key + 1)
-                    break
-            else:
-                node_keys = list(node.keys())
-                if key != node_keys[-1]:
-                    next_key = node_keys[node_keys.index(key)+1]
-                    end_row, end_col = node.lc.key(next_key)
-                    break
-        else:
-            end_row = len(self.meta_yaml)
-            end_col = len(self.meta_yaml[-1])
-        return (start_row, start_col, end_row, end_col)
-
-    def get_raw(self, path):
-        """Extracts the unparsed text for a node in the meta.yaml
-
-        Args:
-          path: Slash-separated path to the node. Numbers can be used
-                to access indices in lists. A number '0' is ignored if
-                the node is a dict (so 'source/0/url' will work even if
-                there is only one url).
-
-        Returns:
-          Extracted raw text
-        """
-        start_row, start_col, end_row, end_col = self.get_raw_range(path)
-        lines = []
-        lines.append(self.meta_yaml[start_row][start_col:])
-        for row in range(start_row+1, end_row):
-            lines.append(self.meta_yaml[row])
-        lines.append(self.meta_yaml[end_row][:end_col])
-        return "\n".join(lines).strip()
-
-
 class Filter(abc.ABC):
     """Function object type called by Scanner"""
-    def __init__(self, scanner: "Scanner") -> None:
+    def __init__(self, scanner: "Scanner", *_args, **_kwargs) -> None:
         self.scanner = scanner
 
     @abc.abstractmethod
@@ -603,7 +169,7 @@ class Scanner():
         self.proc_pool_executor = ProcessPoolExecutor(3)
         self.limit_inflight = asyncio.Semaphore(max_inflight)
 
-    def add(self, filt: Filter, *args, **kwargs) -> None:
+    def add(self, filt: Type[Filter], *args, **kwargs) -> None:
         """Adds `Filter` to this `Scanner`"""
         self.filters.append(filt(self, *args, **kwargs))
 
@@ -718,7 +284,8 @@ class Scanner():
     async def get_checksum_from_url(self, url: str, desc: str) -> str:
         """Compute sha256 checksum of content at **url**
 
-        Shows TQDM progress monitor with label **desc**.
+        - Shows TQDM progress monitor with label **desc**.
+        - Caches result
         """
         if self.cache and url in self.cache["url_checksum"]:
             return self.cache["url_checksum"][url]
@@ -737,6 +304,10 @@ class Scanner():
     @backoff.on_exception(backoff.fibo, aiohttp.ClientResponseError, max_tries=20,
                           giveup=lambda ex: ex.code not in [429, 502, 503, 504])
     async def get_checksum_from_http(self, url: str, desc: str) -> str:
+        """Compute sha256 checksum of content at http **url**
+
+        Shows TQDM progress monitor with label **desc**.
+        """
         checksum = sha256()
         async with self.session.get(url) as resp:
             resp.raise_for_status()
@@ -754,6 +325,10 @@ class Scanner():
     @backoff.on_exception(backoff.fibo, aiohttp.ClientResponseError, max_tries=20,
                           giveup=lambda ex: ex.code not in [429, 502, 503, 504])
     async def get_file_from_url(self, fname: str, url: str, desc: str) -> None:
+        """Fetch file at **url** into **fname**
+
+        Shows TQDM progress monitor with label **desc**.
+        """
         async with self.session.get(url) as resp:
             resp.raise_for_status()
             size = int(resp.headers.get("Content-Length", 0))
@@ -768,6 +343,7 @@ class Scanner():
                         progress.update(len(block))
 
     async def get_ftp_listing(self, url):
+        """Returns list of files at FTP **url**"""
         logger.debug("FTP: listing %s", url)
         if self.cache and url in self.cache["ftp_list"]:
             return self.cache["ftp_list"][url]
@@ -775,12 +351,17 @@ class Scanner():
         parsed = urlparse(url)
         async with aioftp.ClientSession(parsed.netloc,
                                         password=self.USER_AGENT+"@") as client:
-            res = [str(path) for path, _info in (await client.list(parsed.path))]
+            res = [str(path) for path, _info in await client.list(parsed.path)]
         if self.cache:
             self.cache["ftp_list"][url] = res
         return res
 
-    async def get_checksum_from_ftp(self, url, desc):
+    async def get_checksum_from_ftp(self, url, _desc=None):
+        """Compute sha256 checksum of content at ftp **url**
+
+        Does not show progress monitor at this time (would need to
+        get file size first)
+        """
         parsed = urlparse(url)
         checksum = sha256()
         async with aioftp.ClientSession(parsed.netloc,
@@ -795,6 +376,8 @@ class ExcludeOtherChannel(Filter):
     """Filters recipes matching packages in other **channels**"""
 
     class OtherChannel(RecipeError):
+        """This recipe builds one or more packages that are also present
+        in at least one other channel"""
         template = "builds package found in other channel(s)"
         level = logging.DEBUG
 
@@ -802,10 +385,10 @@ class ExcludeOtherChannel(Filter):
                  cache: str) -> None:
         super().__init__(scanner)
         logger.info("Loading package lists for %s", channels)
-        r = utils.RepoData()
+        repo = utils.RepoData()
         if cache:
-            r.set_cache(cache)
-        self.other = set(r.get_package_data('name', channels=channels))
+            repo.set_cache(cache)
+        self.other = set(repo.get_package_data('name', channels=channels))
 
     async def apply(self, recipe):
         if any(package in self.other
@@ -822,6 +405,7 @@ class ExcludeSubrecipe(Filter):
     """
 
     class IsSubRecipe(RecipeError):
+        """This recipe is a sub-recipe (i.e. blast/2.4.0)."""
         template = "is a subrecipe"
         level = logging.DEBUG
 
@@ -841,6 +425,7 @@ class ExcludeBlacklisted(Filter):
     """Filters blacklisted recipes"""
 
     class Blacklisted(RecipeError):
+        """This recipe has been blacklisted (fails to build, see blacklist file)"""
         template = "is blacklisted"
         level = logging.DEBUG
 
@@ -876,26 +461,45 @@ class UpdateVersion(Filter):
     """
 
     class Metapackage(RecipeError):
+        """This recipe only builds a meta package - it has no upstream sources."""
         template = "builds a meta package (recipe has no sources)"
 
     class UpToDate(RecipeError):
+        """This recipe appears to be up to date!"""
         template = "is up to date"
         level = logging.DEBUG
 
     class UpdateVersionFailure(RecipeError):
+        """This recipe did not show the expected version number after updating.
+
+        Something probably went wrong trying to replace the version number in the
+        meta.yaml.
+        """
         template = "could not be updated from %s to %s"
 
     class NoUrlInSource(RecipeError):
+        """This recipe has a source without an URL.
+
+        Most likely, this is due to a git-url.
+        """
         template = "has no URL in source %i"
 
     class NoRecognizedSourceUrl(RecipeError):
+        """None of the source URLs in this recipe were recognized."""
         template = "has no URL in source %i recognized by any Hoster class"
         level = logging.DEBUG
 
     class UrlNotVersioned(RecipeError):
+        """This recipe has an URL unaffected by the version change.
+        the recipe"""
         template = "has URL not modified by version change"
 
     class NoReleases(RecipeError):
+        """No releases at all were found for this recipe.
+
+        This is unusual - something probably went wrong finding releases for
+        the source urls in this recipe.
+        """
         template = "has no releases?!"
 
     def __init__(self, scanner: "Scanner", hoster_factory,
@@ -948,8 +552,8 @@ class UpdateVersion(Filter):
                 raise self.UpToDate(recipe)
 
         # Update `url:`s without Jinja expressions (plain text)
-        for fn in versions[latest]:
-            recipe.replace(fn, versions[latest][fn]['link'], within=["source"])
+        for fname in versions[latest]:
+            recipe.replace(fname, versions[latest][fname]['link'], within=["source"])
 
         # Update the version number itself. This will also usually update
         # `url:`s expressed with `{{version}}` tags.
@@ -962,7 +566,7 @@ class UpdateVersion(Filter):
         recipe.render()
 
         # Verify that the rendered recipe has the right version number
-        if not VersionOrder(recipe.version) == VersionOrder(latest):
+        if VersionOrder(recipe.version) != VersionOrder(latest):
             raise self.UpdateVersionFailure(recipe, recipe.orig.version, latest)
 
         # Verify that every url was modified
@@ -988,8 +592,8 @@ class UpdateVersion(Filter):
             for num, source in enumerate(source_iter):
                 add_versions = await self.get_versions(recipe, source, num+1)
                 for vers, files in add_versions.items():
-                    for fn, data in files.items():
-                        versions[vers][fn] = data
+                    for fname, data in files.items():
+                        versions[vers][fname] = data
         else:
             versions = await self.get_versions(recipe, sources, 0)
 
@@ -1063,7 +667,7 @@ class UpdateVersion(Filter):
         return latest
 
     def check_version_pin_conflict(self, recipe: Recipe,
-                                   versions: List[Dict[str, Any]]) -> List[str]:
+                                   versions: Dict[str, Any]) -> None:
         """Find items in **versions** conflicting with pins
 
         Example:
@@ -1078,20 +682,20 @@ class UpdateVersion(Filter):
             for pkg, spec in depends:
                 norm_pkg = pkg.replace("-", "_")
                 try:
-                    ms = MatchSpec(version=spec.replace(" ",""))
+                    mspec = MatchSpec(version=spec.replace(" ", ""))
                 except conda.exceptions.InvalidVersionSpecError:
                     logger.error("Recipe %s: invalid upstream spec %s %s",
                                  recipe, pkg, repr(spec))
                     continue
                 if norm_pkg in variants[0]:
                     for variant in variants:
-                        if not ms.match({'name': '', 'build': '', 'build_number': '0',
-                                         'version': variant[norm_pkg]}):
+                        if not mspec.match({'name': '', 'build': '', 'build_number': '0',
+                                            'version': variant[norm_pkg]}):
                             logger.error("Recipe %s: %s %s conflicts pins",
                                          recipe, pkg, spec)
 
-        for version, files in versions.items():
-            for fn, data in files.items():
+        for files in versions.values():
+            for data in files.values():
                 depends = data.get('depends')
                 if depends:
                     check_pins(depends.items())
@@ -1127,15 +731,19 @@ class UpdateChecksums(Filter):
     """Download upstream source files, recompute checksum and update Recipe"""
 
     class NoValidUrls(RecipeError):
+        """Failed to download any file for a source to generate new checksum"""
         template = "has no valid urls in source %i"
 
     class SourceUrlMismatch(RecipeError):
+        """The URLs in a source point to different files after update"""
         template = "has urls in source %i pointing to different files"
 
     class ChecksumReplaceFailed(RecipeError):
+        """The checksum could not be updated after version bump"""
         template = ": failed to replace checksum"
 
     class ChecksumUnchanged(RecipeError):
+        """The checksum did not change after version bump"""
         template = "had no change to checksum after update?!"
 
     def __init__(self, scanner: "Scanner",
@@ -1159,7 +767,7 @@ class UpdateChecksums(Filter):
         logger.info("Updating checksum for %s %s", recipe, recipe.version)
         if isinstance(sources, Mapping):
             sources = [sources]
-        for source_idx,  source in enumerate(sources):
+        for source_idx, source in enumerate(sources):
             await self.update_source(recipe, source, source_idx)
         recipe.render()
         return recipe
@@ -1248,7 +856,14 @@ class GitFilter(Filter):
 
 
 class ExcludeNoActiveUpdate(GitFilter):
+    """Allows only recipes that already have an update active
+
+    Requires that the recipes are loaded from git - recipes with active
+    updates are those for which a branch with commits newer than master
+    exists (which is checked by GitLoadRecipe).
+    """
     class NoActiveUpdate(RecipeError):
+        """Recipe is not currently being updated"""
         template = "is not currently being updated"
         level = logging.DEBUG
 
@@ -1266,8 +881,9 @@ class LoadRecipe(Filter):
         self.sem = asyncio.Semaphore(10)
 
     async def apply(self, recipe):
-        async with self.sem, aiofiles.open(recipe.path, encoding="utf-8") as fd:
-            recipe_text = await fd.read()
+        async with self.sem, \
+                   aiofiles.open(recipe.path, encoding="utf-8") as fdes:
+            recipe_text = await fdes.read()
         recipe = await self.scanner.run_sp(recipe.load_from_string, recipe_text)
         recipe.set_original()
         return recipe
@@ -1327,8 +943,9 @@ class WriteRecipe(Filter):
         self.sem = asyncio.Semaphore(10)
 
     async def apply(self, recipe):
-        async with self.sem, aiofiles.open(recipe.path, "w", encoding="utf-8") as fd:
-            await fd.write(recipe.dump())
+        async with self.sem, \
+                   aiofiles.open(recipe.path, "w", encoding="utf-8") as fdes:
+            await fdes.write(recipe.dump())
         return recipe
 
 
@@ -1336,16 +953,17 @@ class GitWriteRecipe(GitFilter):
     """Writes `Recipe` to git repo"""
 
     class NoChanges(RecipeError):
+        """Recipe had no changes after applying updates"""
         template = "had no changes"
 
     async def apply(self, recipe):
         branch_name = self.branch_name(recipe)
-        remote_branch = self.git.get_remote_branch(branch_name)
         changed = False
         async with self.git.lock_working_dir:
             self.git.prepare_branch(branch_name)
-            async with aiofiles.open(recipe.path, "w", encoding="utf-8") as fd:
-                await fd.write(recipe.dump())
+            async with aiofiles.open(recipe.path, "w",
+                                     encoding="utf-8") as fdes:
+                await fdes.write(recipe.dump())
             changed = self.git.commit_and_push_changes(recipe, branch_name)
         if changed:
             # CircleCI appears to have problems picking up on our PRs. Let's wait
@@ -1360,12 +978,21 @@ class CreatePullRequest(GitFilter):
     """Creates or Updates PR on GitHub"""
 
     class UpdateInProgress(RecipeError):
+        """The recipe has an active update PR"""
         template = "has update in progress"
 
     class UpdateRejected(RecipeError):
+        """This update has been rejected previously on Github
+
+        A PR created by Autobump for this recipe and this version
+        already exists on Github and has been closed. We take this as
+        indication that the update should not be repeated
+
+        """
         template = "had latest updated rejected manually"
 
     class FailedToCreatePR(RecipeError):
+        """Something went wrong trying to create a new PR on Github."""
         template = "had failure in PR create"
 
     def __init__(self, scanner, git_handler, token: str,
@@ -1375,29 +1002,42 @@ class CreatePullRequest(GitFilter):
         super().__init__(scanner, git_handler)
         self.to_user = to_user
         self.to_repo = to_repo
-        self.gh: GitHubHandler = AiohttpGitHubHandler(
+        self.ghub: GitHubHandler = AiohttpGitHubHandler(
             token, dry_run, to_user, to_repo)
 
     async def async_init(self):
         """Create gidget GithubAPI object from session"""
-        await self.gh.login(self.scanner.session, self.scanner.USER_AGENT)
+        await self.ghub.login(self.scanner.session, self.scanner.USER_AGENT)
         await asyncio.sleep(1)  # let API settle
 
-    def get_deps_diff(self, recipe):
-        diffset = {'host': set(), 'run': set()}
+    @staticmethod
+    def render_deps_diff(recipe):
+        """Renders a "diff" of the recipes upstream dependencies.
+
+        This relies on the 'depends' data structure in the recipe's
+        ``version_data`. This structure is expected to be a dict with
+        two keys 'host' and 'run', each of which contain dict of
+        package to version dependency mappings. The 'depends' data can
+        be created by the **hoster** (currently done by CPAN and CRAN)
+        or filled in later by the `FetchUpstreamDependcies` filter
+        (currently for PyPi).
+        """
+        diffset: Dict[str, Set[str]] = {'host': set(), 'run': set()}
         if not recipe.version_data:
             logger.debug("Recipe %s: dependency diff not rendered (no version_data)", recipe)
-        for fn in recipe.version_data:
-            if fn not in recipe.orig.version_data:
-                logger.debug("Recipe %s: dependency diff not rendered (no orig.version_data)", recipe)
+        for fname in recipe.version_data:
+            if fname not in recipe.orig.version_data:
+                logger.debug("Recipe %s: dependency diff not rendered (no orig.version_data)",
+                             recipe)
                 continue
-            new = recipe.version_data[fn].get('depends')
-            orig = recipe.orig.version_data[fn].get('depends')
+            new = recipe.version_data[fname].get('depends')
+            orig = recipe.orig.version_data[fname].get('depends')
             if not new or not orig:
-                logger.debug("Recipe %s: dependency diff not rendered (no depends in version_data)", recipe)
+                logger.debug("Recipe %s: dependency diff not rendered (no depends in version_data)",
+                             recipe)
                 continue
             for kind in ('host', 'run'):
-                deps = set()
+                deps: Set[str] = set()
                 deps.update(new[kind].keys())
                 deps.update(orig[kind].keys())
                 for dep in deps:
@@ -1406,9 +1046,9 @@ class CreatePullRequest(GitFilter):
                     elif dep not in orig[kind]:
                         diffset[kind].add("+   - {} {}".format(dep, new[kind][dep]))
                     elif orig[kind][dep] != new[kind][dep]:
-                        diffset[kind].add("-   - {} {}\n"
-                                          "+   - {} {}".format(dep, orig[kind][dep],
-                                                               dep, new[kind][dep]))
+                        diffset[kind].add("-   - {dep} {}\n"
+                                          "+   - {dep} {}".format(orig[kind][dep], new[kind][dep],
+                                                                  dep=dep))
         text = ""
         for kind, lines in diffset.items():
             if lines:
@@ -1418,73 +1058,81 @@ class CreatePullRequest(GitFilter):
             logger.debug("Recipe %s: dependency diff not rendered (all good)", recipe)
         return text
 
+    @staticmethod
+    def get_github_author(recipe) -> Optional[str]:
+        """Fetches upstream Github account name if applicable
+
+        For recipes with upstream sources hosted on Github, we can extract
+        the account name from the source URL, so that we can add an @mention
+        notifying the upstream author of the update.
+        """
+        for ver in recipe.version_data.values():
+            if 'hoster' in ver and ver['hoster'].__class__.__name__.startswith('Github'):
+                return ver['vals']['account']
+        return None
+
     async def apply(self, recipe):
         branch_name = self.branch_name(recipe)
         template = utils.jinja.get_template("bump_pr.md")
-        author = None
-        for v in recipe.version_data.values():
-            if 'hoster' in v and v['hoster'].__class__.__name__.startswith('Github'):
-                author = v['vals']['account']
-                break
+        author = self.get_github_author(recipe)
 
         body = template.render({
             'r': recipe,
             'recipe_relurl': "/".join((self.to_user, self.to_repo, 'tree',
                                        branch_name, recipe.basedir + recipe.reldir)),
             'author': author,
-            'author_is_member': await self.gh.is_member(author),
-            'dependency_diff': self.get_deps_diff(recipe),
+            'author_is_member': await self.ghub.is_member(author),
+            'dependency_diff': self.render_deps_diff(recipe),
         })
         labels = ["autobump"]
         title = f"Update {recipe} to {recipe.version}"
 
         # check if we already have an open PR (=> update in progress)
-        prs = await self.gh.get_prs(from_branch=branch_name, from_user="bioconda")
-        if prs:
-            if len(prs) > 1:
+        pullreqs = await self.ghub.get_prs(from_branch=branch_name, from_user="bioconda")
+        if pullreqs:
+            if len(pullreqs) > 1:
                 logger.error("Multiple PRs updating %s: %s",
                              recipe,
-                             ", ".join(str(pr['number']) for pr in prs))
-            for pr in prs:
+                             ", ".join(str(pull['number']) for pull in pullreqs))
+            for pull in pullreqs:
                 logger.debug("Found PR %i updating %s: %s",
-                             pr["number"], recipe, pr["title"])
+                             pull["number"], recipe, pull["title"])
             # update the PR if title or body changed
-            pr = prs[0]
+            pull = pullreqs[0]
             args = {}
-            if body != pr["body"]:
+            if body != pull["body"]:
                 args["body"] = body
-            if title != pr["title"]:
+            if title != pull["title"]:
                 args["title"] = title
             if args:
-                prs = await self.gh.modify_issue(number=pr['number'], **args)
-                if prs:
+                if await self.ghub.modify_issue(number=pull['number'], **args):
                     logger.info("Updated PR %i updating %s to %s",
-                                pr['number'], recipe, recipe.version)
+                                pull['number'], recipe, recipe.version)
                 else:
                     logger.error("Failed to update PR %i with %s",
-                                 pr['number'], args)
+                                 pull['number'], args)
             else:
                 logger.debug("Not updating PR %i updating %s - no changes",
-                             pr['number'], recipe)
+                             pull['number'], recipe)
 
             raise self.UpdateInProgress(recipe)
 
         # check for matching closed PR (=> update rejected)
-        prs = await self.gh.get_prs(from_branch=branch_name, state=self.gh.STATE.closed)
-        if any(recipe.version in pr['title'] for pr in prs):
+        pullreqs = await self.ghub.get_prs(from_branch=branch_name, state=self.ghub.STATE.closed)
+        if any(recipe.version in pull['title'] for pull in pullreqs):
             raise self.UpdateRejected(recipe)
 
         # finally, create new PR
-        pr = await self.gh.create_pr(title=title,
-                                     from_branch=branch_name,
-                                     body=body)
-        if not pr:
+        pull = await self.ghub.create_pr(title=title,
+                                         from_branch=branch_name,
+                                         body=body)
+        if not pull:
             raise self.FailedToCreatePR(recipe)
 
         # set labels (can't do that in create_pr as they are part of issue API)
-        await self.gh.modify_issue(number=pr['number'], labels=labels)
+        await self.ghub.modify_issue(number=pull['number'], labels=labels)
 
-        logger.info("Created PR %i updating %s to %s", pr['number'], recipe, recipe.version)
+        logger.info("Created PR %i updating %s to %s", pull['number'], recipe, recipe.version)
         return recipe
